requirements.txt
asgiref==3.7.2
Django==3.2
django-ckeditor==6.6.1
django-cleanup==8.0.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.17
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
